deleted the wrong main.py, kept the one in src/
