+++ src/main.py
@@ -1,10 +1,12 @@
 import sys
 import random
+import os
 from PySide6 import QtCore, QtWidgets, QtGui
 from PySide6.QtCore import Signal
+from PIL import Image
 
 class MyWidget(QtWidgets.QWidget):
-   def __init__(self, parent):
+   def __init__(self, image_loaded):
       super().__init__()
 
       self.hello = ["Hallo Welt", "Hei maailma", "Hola Mundo", "Привет мир", "Hewwo Wowwd", "Hello World"]
@@ -19,18 +21,17 @@ class MyWidget(QtWidgets.QWidget):
       self.layout.addWidget(self.button)
 
       self.button.clicked.connect(self.magic)
-      parent.image_loaded.connect(self.show_image)
+
+      scene = QtWidgets.QGraphicsScene()
+      scene.setSceneRect(0, 0, 200, 200)
+      text = scene.addText("Hello!")
+
+      view = MyView(image_loaded)
+      self.layout.addWidget(view)
 
    @QtCore.Slot()
    def magic(self):
       self.label.setText(random.choice(self.hello))
-
-   @QtCore.Slot()
-   def show_image(self, file_name):
-      label = QtWidgets.QLabel(self)
-      pixmap = QtGui.QPixmap(file_name)
-      label.setPixmap(pixmap)
-      self.layout.addWidget(label)
 
 class MainWindow(QtWidgets.QMainWindow):
    image_loaded = Signal(str)
@@ -39,7 +40,7 @@ class MainWindow(QtWidgets.QMainWindow):
       super().__init__()
 
       self.setWindowTitle("BeadPro")
-      main_widget = MyWidget(self)
+      main_widget = MyWidget(self.image_loaded)
       self.setCentralWidget(main_widget)
 
       open_icon = QtGui.QIcon.fromTheme("document-open", QtGui.QIcon(":/resources/folder_open.png"))
@@ -53,7 +54,8 @@ class MainWindow(QtWidgets.QMainWindow):
    @QtCore.Slot()
    def open_image(self):
       file_name = QtWidgets.QFileDialog.getOpenFileName(self, "Open Image", "C:/", "Image Files (*.png)")
-      print(file_name)
+      print("Inside open_image")
+      print(file_name[0])
       self.image_loaded.emit(file_name[0])
 
 def window():
@@ -64,6 +66,34 @@ def window():
    window.show()
 
    sys.exit(app.exec())
+
+class MyView(QtWidgets.QGraphicsView):
+   def __init__(self, image_loaded):
+      QtWidgets.QGraphicsView.__init__(self)
+      self.setGeometry(QtCore.QRect(100, 100, 600, 250))
+      self.scene = QtWidgets.QGraphicsScene(self)
+      self.scene.setSceneRect(QtCore.QRectF())
+      self.setScene(self.scene)
+      image_loaded.connect(self.draw_image)
+
+   @QtCore.Slot()
+   def draw_image(self, file_name):
+      print("Inside draw_image")
+      print(file_name)
+      im = Image.open(file_name)
+
+      pixels = list(im.getdata())
+      width, height = im.size
+      pixels = [pixels[i * width:(i + 1) * width] for i in range(height)]
+
+      WIDTH = 30
+      HEIGHT = 30
+      pen = QtGui.QPen()
+      pen.setStyle(QtCore.Qt.PenStyle.NoPen)
+      for i in range(len(pixels)):
+         for j in range(len(pixels[0])):
+            rgba = pixels[i][j]
+            self.scene.addRect(j*WIDTH, i*HEIGHT, WIDTH, HEIGHT, pen, QtGui.QBrush(QtGui.QColor(rgba[0], rgba[1], rgba[2])))
 
 if __name__ == '__main__':
    window()
